refactor: OpenApiPagingItemReader 에서 UrlProvider 제거
