+++ src/main/java/com/foodiefinder/datapipeline/reader/OpenApiPagingItemReader.java
@@ -1,28 +1,21 @@
 package com.foodiefinder.datapipeline.reader;
 
 import com.foodiefinder.datapipeline.enums.JobState;
-import com.foodiefinder.datapipeline.provider.UrlProvider;
-import com.foodiefinder.datapipeline.util.HttpRequest;
+import com.foodiefinder.datapipeline.util.RequestStrategy;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 
-public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
+public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I> {
 
-    private UrlProvider<String> urlProvider;
-    private HttpRequest<I> httpRequest;
-
-    /**
-     * Java에서 제네릭 타입의 클래스 정보를 얻기 위해서는 리플렉션을 사용해야
-     * 하지만 자바의 타입 소거 때문에 런타임에 제네릭 타입의 정보가 손실됩니다.
-     * 다음과 같은 이유로 생성
-     */
+    private RequestStrategy<I, Map<String, String>> request;
     private Class<I> itemType;
 
-    private I endOfPageBody;
+    private String endOfPageBody;
     private HttpStatusCode endOfPageHttpStatus;
 
     private List<String> apiUrlList;
@@ -31,10 +24,7 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
 
     @Override
     protected I doRead() {
-        // Parameter 로 Url 을 만들고 요청을 한다.
-        String url = urlProvider.getUrl(apiUrlList.get(apiUrlListIndex), getParams());
-        ResponseEntity<I> responseEntity = httpRequest.sendRequest(url, itemType);
-
+        ResponseEntity<I> responseEntity = request.sendRequest(apiUrlList.get(apiUrlListIndex), getParams(), itemType);
         I body = responseEntity.getBody();
         HttpStatusCode statusCode = responseEntity.getStatusCode();
 
@@ -47,8 +37,7 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
                 getParams().put(getPageName(), String.valueOf(getPageStartIndex()));
                 return null;
             }
-        }
-        else{
+        } else {
             // 마지막 페이지 아님 pageIndex 증가
             return body;
         }
@@ -63,28 +52,23 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
     @Override
     public void open() {
         super.open();
-        // 재시도 로직인경우 url 리스트 인덱스 변경, 재시도 횟수 초과후에는 다음 url로 변경
         Optional<Boolean> optionalRetry = getStateHandler().loadState(JobState.RETRY.name(), Boolean.class);
         Optional<Boolean> optionalNext = getStateHandler().loadState(JobState.NEXT.name(), Boolean.class);
 
-        if(optionalRetry.isPresent() && Boolean.TRUE.equals(optionalRetry.get())) {
+        if (optionalRetry.isPresent() && Boolean.TRUE.equals(optionalRetry.get())) {
             apiUrlListIndex = getStateHandler().loadState("apiUrlListIndex", Integer.class)
                     .orElse(0);
-        }
-        else if (optionalNext.isPresent() && Boolean.TRUE.equals(optionalNext.get())) {
+        } else if (optionalNext.isPresent() && Boolean.TRUE.equals(optionalNext.get())) {
             apiUrlListIndex = getStateHandler().loadState("apiUrlListIndex", Integer.class)
                     .orElse(0) + 1;
         }
-
-        httpRequest = new HttpRequest<>();
     }
 
     @Override
     public void update() {
         super.update();
-        // 한번의 리드가 끝나면 상태 저장
         getStateHandler().saveState(JobState.NEXT.name(), false);
-        getStateHandler().saveState("apiUrlListIndex",apiUrlListIndex);
+        getStateHandler().saveState("apiUrlListIndex", apiUrlListIndex);
     }
 
     @Override
@@ -93,9 +77,8 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
         apiUrlListIndex = 0;
     }
 
-    // 특정 응답을 받으면 페이지의 끝 임을 인지하도록
     private boolean isAtEndOfPage(I body, HttpStatusCode httpStatus) {
-        if (endOfPageHttpStatus.equals(httpStatus) && endOfPageBody.equals(body)){
+        if (endOfPageHttpStatus.equals(httpStatus) && endOfPageBody.equals(body)) {
             return true;
         }
         return false;
@@ -105,19 +88,9 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
         return apiUrlList.size() <= apiUrlListIndex;
     }
 
-    /**
-     * 마지막 페이지의 body, status
-     * @param body
-     * @param httpStatus
-     */
-    public void setEndOfPageResponse(I body, HttpStatusCode httpStatus) {
+    public void setEndOfPageResponse(String body, HttpStatusCode httpStatus) {
         this.endOfPageBody = body;
         this.endOfPageHttpStatus = httpStatus;
-    }
-
-    // 페이징을 할 url 생성자
-    public void setUrlProvider(UrlProvider urlProvider) {
-        this.urlProvider = urlProvider;
     }
 
     public void setApiUrlList(List<String> apiUrlList) {
@@ -126,5 +99,9 @@ public class OpenApiPagingItemReader<I> extends AbstractPagingItemReader<I>{
 
     public void setItemType(Class<I> itemType) {
         this.itemType = itemType;
+    }
+
+    public void setRequest(RequestStrategy<I, Map<String, String>> request) {
+        this.request = request;
     }
 }
