+++ src/main/java/com/foodiefinder/datapipeline/util/UrlProvider.java
@@ -1,4 +1,4 @@
-package com.foodiefinder.datapipeline.provider;
+package com.foodiefinder.datapipeline.util;
 
 import java.util.Map;
