+++ src/main/java/com/foodiefinder/datapipeline/util/UrlWithParamsProvider.java
@@ -1,4 +1,4 @@
-package com.foodiefinder.datapipeline.provider;
+package com.foodiefinder.datapipeline.util;
 
 import org.springframework.web.util.UriComponentsBuilder;
